Made game interactive
The buttons now are functional. It takes the answer of the button you pick and  the random number for the computer. If the player and computer have the same pick, it doesn't count towards score. If whoever wins, the score count to them.
The color theme of the page is changed to darker.